Refactor to torch_utils
