+++ tools/rocm_utils.py
@@ -33,7 +33,7 @@ def install_torch_deps():
 
 
 def install_pytorch_nightly(rocm_version: str, env, dryrun=False):
-    from .torch.utils import TORCH_NIGHTLY_PACKAGES
+    from .torch_utils import TORCH_NIGHTLY_PACKAGES
     uninstall_torch_cmd = ["pip", "uninstall", "-y"]
     uninstall_torch_cmd.extend(TORCH_NIGHTLY_PACKAGES)
     if dryrun:
